% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_prep_dataframe.R
\name{pcm_prep_dataframe}
\alias{pcm_prep_dataframe}
\title{Prepare dataframe for efficient use of PC Miler API}
\usage{
pcm_prep_dataframe(df)
}
\arguments{
\item{df}{A dataframe that contains both origin and destination columns.}
}
\value{
A list containing prepared data for processing via PC Miler API.
}
\description{
Prepare dataframe for efficient use of PC Miler API
}
\examples{
pcm_prep_dataframe(df_route_samples)
}
