% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_Route_json.R
\name{pcm_Route_json}
\alias{pcm_Route_json}
\title{Convert dataframe of origin/destination pairs to JSON file for PC Miler API}
\usage{
pcm_Route_json(
  df,
  ReportType = "MileageReportType",
  lst_origins = lst_origins,
  lst_destinations = lst_destinations
)
}
\arguments{
\item{df}{A dataframe containing origin/destination column(s).}

\item{lst_destinations}{}
}
\description{
Convert dataframe of origin/destination pairs to JSON file for PC Miler API
}
