% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_df_to_size.R
\name{split_df_to_size}
\alias{split_df_to_size}
\title{Create list of dataframes from original dataframe with max rows_per_chunk}
\usage{
split_df_to_size(df, rows_per_chunk = 20L)
}
\arguments{
\item{df}{A dataframe.}

\item{rows_per_chunk}{An integer indicating the max number of rows for each dataframe in the resulting list.}
}
\value{
A list of dataframes each sized to 'rows_per_chunk' or less.
}
\description{
Create list of dataframes from original dataframe with max rows_per_chunk
}
\examples{
split_df_to_size(df_route_samples)
split_df_to_size(df_route_samples, rows_per_chunk = 13)
}
